server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0
pymongo==4.3.3
pyjwt==2.6.0
werkzeug==2.2.3